# Add new randomized benchmarks