--- a/src/benchmarks/bench.cpp
+++ b/src/benchmarks/bench.cpp
@@ -337,7 +337,6 @@ void runMatchBenchmark(const OSRM &osrm, const GPSTraces &gpsTraces)
     };
 
     std::vector<Benchmark> benchmarks = {{"1000 matches, default radius"},
-                                         {"1000 matches, default radius2"},
                                          {"1000 matches, radius=10", 10},
                                          {"1000 matches, radius=20", 20}};