--- a/src/benchmarks/route.cpp
+++ b/src/benchmarks/route.cpp
@@ -164,4 +164,3 @@ catch (const std::exception &e)
     std::cerr << "Error: " << e.what() << std::endl;
     return EXIT_FAILURE;
 }
-