The goal here is to have more "diversified" benchmarks, i.e.:
1. Be able to run it against bigger dataset than just Berlin (it is enabled via adding magic "Performance" label to PR): I chose Poland as it was small enough to somehow work on CI machines from GitHub :)
2. Be able to cover more cases via using random routes, but not fixed ones

It is kind of similar to e2e tests we added in one of recent PRs, but benchmarks are written in C++ and run against libosrm, but not osrm-routed.




<!-- BENCHMARK_RESULTS_START -->
<details><summary><h2>Benchmark Results</h2></summary>

| Benchmark | Base | PR |
|-----------|------|----|
| alias | aliased u32: 1083.64<br/>plain u32: 1079.6<br/>aliased double: 953.734<br/>plain double: 950.187 | aliased u32: 1093.13<br/>plain u32: 1093.32<br/>aliased double: 953.474<br/>plain double: 947.916 |
| e2e_match_ch | Total: 57224.499225616455ms<br/>Min time: 4.024982452392578ms<br/>Mean time: 59.11621820828146ms<br/>Median time: 41.547536849975586ms<br/>95th percentile: 183.1827044486999ms<br/>99th percentile: 434.13395643234264ms<br/>Max time: 492.4740791320801ms | Total: 54882.668018341064ms<br/>Min time: 4.048824310302734ms<br/>Mean time: 56.69697109332755ms<br/>Median time: 40.22789001464844ms<br/>95th percentile: 168.15186738967896ms<br/>99th percentile: 398.0480027198792ms<br/>Max time: 458.0681324005127ms |
| e2e_match_mld | Total: 38384.233713150024ms<br/>Min time: 3.398895263671875ms<br/>Mean time: 39.65313400118804ms<br/>Median time: 30.88390827178955ms<br/>95th percentile: 119.09809112548822ms<br/>99th percentile: 253.042521476746ms<br/>Max time: 296.2648868560791ms | Total: 38459.699630737305ms<br/>Min time: 3.421306610107422ms<br/>Mean time: 39.73109465985259ms<br/>Median time: 30.67755699157715ms<br/>95th percentile: 119.7221755981445ms<br/>99th percentile: 255.00832557678245ms<br/>Max time: 299.21770095825195ms |
| e2e_nearest_ch | Total: 1356.0309410095215ms<br/>Min time: 1.1560916900634766ms<br/>Mean time: 1.3560309410095215ms<br/>Median time: 1.2714862823486328ms<br/>95th percentile: 1.759326457977295ms<br/>99th percentile: 1.7977166175842285ms<br/>Max time: 2.1240711212158203ms | Total: 1356.16135597229ms<br/>Min time: 1.1622905731201172ms<br/>Mean time: 1.35616135597229ms<br/>Median time: 1.270890235900879ms<br/>95th percentile: 1.7720460891723633ms<br/>99th percentile: 1.8274998664855957ms<br/>Max time: 1.8966197967529297ms |
| e2e_nearest_mld | Total: 1356.9645881652832ms<br/>Min time: 1.1353492736816406ms<br/>Mean time: 1.3569645881652832ms<br/>Median time: 1.2695789337158203ms<br/>95th percentile: 1.7695426940917969ms<br/>99th percentile: 1.8039345741271973ms<br/>Max time: 2.173900604248047ms | Total: 1338.1531238555908ms<br/>Min time: 1.1463165283203125ms<br/>Mean time: 1.3381531238555908ms<br/>Median time: 1.2562274932861328ms<br/>95th percentile: 1.7443299293518066ms<br/>99th percentile: 1.79368257522583ms<br/>Max time: 1.9650459289550781ms |
| e2e_route_ch | Total: 4331.030607223511ms<br/>Min time: 1.3818740844726562ms<br/>Mean time: 4.331030607223511ms<br/>Median time: 4.546999931335449ms<br/>95th percentile: 6.4900994300842285ms<br/>99th percentile: 7.155919075012207ms<br/>Max time: 8.391857147216797ms | Total: 4376.399040222168ms<br/>Min time: 1.4405250549316406ms<br/>Mean time: 4.376399040222168ms<br/>Median time: 4.576206207275391ms<br/>95th percentile: 6.480705738067626ms<br/>99th percentile: 7.098591327667236ms<br/>Max time: 7.613658905029297ms |
| e2e_route_mld | Total: 6212.393283843994ms<br/>Min time: 1.4238357543945312ms<br/>Mean time: 6.212393283843994ms<br/>Median time: 6.524801254272461ms<br/>95th percentile: 9.512269496917725ms<br/>99th percentile: 10.237679481506348ms<br/>Max time: 10.898113250732422ms | Total: 6121.039867401123ms<br/>Min time: 1.4231204986572266ms<br/>Mean time: 6.121039867401123ms<br/>Median time: 6.423234939575195ms<br/>95th percentile: 9.348511695861816ms<br/>99th percentile: 9.903013706207275ms<br/>Max time: 10.885953903198242ms |
| e2e_table_ch | Total: 18127.747297286987ms<br/>Min time: 2.0105838775634766ms<br/>Mean time: 18.127747297286987ms<br/>Median time: 17.17674732208252ms<br/>95th percentile: 34.09578800201415ms<br/>99th percentile: 36.0596227645874ms<br/>Max time: 37.95123100280762ms | Total: 19232.73515701294ms<br/>Min time: 2.2161006927490234ms<br/>Mean time: 19.23273515701294ms<br/>Median time: 18.1657075881958ms<br/>95th percentile: 36.10975742340088ms<br/>99th percentile: 38.077383041381836ms<br/>Max time: 39.65425491333008ms |
| e2e_table_mld | Total: 210515.6626701355ms<br/>Min time: 10.004043579101562ms<br/>Mean time: 210.5156626701355ms<br/>Median time: 199.46730136871338ms<br/>95th percentile: 408.1838369369507ms<br/>99th percentile: 431.2767624855041ms<br/>Max time: 505.50103187561035ms | Total: 211779.85954284668ms<br/>Min time: 9.959936141967773ms<br/>Mean time: 211.77985954284668ms<br/>Median time: 200.26063919067383ms<br/>95th percentile: 411.81390285491943ms<br/>99th percentile: 436.5763211250305ms<br/>Max time: 464.2190933227539ms |
| e2e_trip_ch | Total: 16362.309455871582ms<br/>Min time: 2.31170654296875ms<br/>Mean time: 16.362309455871582ms<br/>Median time: 16.530275344848633ms<br/>95th percentile: 25.089335441589355ms<br/>99th percentile: 26.507506370544434ms<br/>Max time: 34.74164009094238ms | Total: 17066.81752204895ms<br/>Min time: 2.393960952758789ms<br/>Mean time: 17.06681752204895ms<br/>Median time: 17.2579288482666ms<br/>95th percentile: 26.066029071807858ms<br/>99th percentile: 27.829279899597168ms<br/>Max time: 30.863523483276367ms |
| e2e_trip_mld | Total: 45663.48075866699ms<br/>Min time: 6.854772567749023ms<br/>Mean time: 45.66348075866699ms<br/>Median time: 46.689510345458984ms<br/>95th percentile: 68.59872341156006ms<br/>99th percentile: 70.54884195327759ms<br/>Max time: 73.65894317626953ms | Total: 46039.14475440979ms<br/>Min time: 6.834983825683594ms<br/>Mean time: 46.03914475440979ms<br/>Median time: 46.92816734313965ms<br/>95th percentile: 69.15066242218018ms<br/>99th percentile: 71.47207021713257ms<br/>Max time: 76.27081871032715ms |
| json-render | String: 6.76086ms<br/>Stringstream: 8.81286ms<br/>Vector: 6.84986ms | String: 6.63169ms<br/>Stringstream: 9.40745ms<br/>Vector: 6.95798ms |
| match_ch | Default radius: <br/>4.42754ms/req at 82 coordinate<br/>0.0539944ms/coordinate<br/>Radius 5m: <br/>4.38611ms/req at 82 coordinate<br/>0.0534892ms/coordinate<br/>Radius 10m: <br/>14.9625ms/req at 82 coordinate<br/>0.18247ms/coordinate<br/>Radius 15m: <br/>36.6464ms/req at 82 coordinate<br/>0.446907ms/coordinate<br/>Radius 30m: <br/>313.341ms/req at 82 coordinate<br/>3.82123ms/coordinate | Default radius: <br/>4.41412ms/req at 82 coordinate<br/>0.0538307ms/coordinate<br/>Radius 5m: <br/>4.40452ms/req at 82 coordinate<br/>0.0537136ms/coordinate<br/>Radius 10m: <br/>15.0524ms/req at 82 coordinate<br/>0.183566ms/coordinate<br/>Radius 15m: <br/>36.6571ms/req at 82 coordinate<br/>0.447038ms/coordinate<br/>Radius 30m: <br/>311.97ms/req at 82 coordinate<br/>3.80451ms/coordinate |
| match_mld | Default radius: <br/>2.78561ms/req at 82 coordinate<br/>0.0339709ms/coordinate<br/>Radius 5m: <br/>2.76198ms/req at 82 coordinate<br/>0.0336826ms/coordinate<br/>Radius 10m: <br/>10.2232ms/req at 82 coordinate<br/>0.124673ms/coordinate<br/>Radius 15m: <br/>25.8556ms/req at 82 coordinate<br/>0.315313ms/coordinate<br/>Radius 30m: <br/>303.518ms/req at 82 coordinate<br/>3.70144ms/coordinate | Default radius: <br/>2.77916ms/req at 82 coordinate<br/>0.0338922ms/coordinate<br/>Radius 5m: <br/>2.76928ms/req at 82 coordinate<br/>0.0337717ms/coordinate<br/>Radius 10m: <br/>10.1692ms/req at 82 coordinate<br/>0.124015ms/coordinate<br/>Radius 15m: <br/>26.0818ms/req at 82 coordinate<br/>0.31807ms/coordinate<br/>Radius 30m: <br/>306.859ms/req at 82 coordinate<br/>3.74218ms/coordinate |
| osrm_contract | Time: 827.40s Peak RAM: 2967.22MB | Time: 830.42s Peak RAM: 2963.85MB |
| osrm_customize | Time: 24.16s Peak RAM: 1943.72MB | Time: 24.15s Peak RAM: 1925.04MB |
| osrm_extract | Time: 274.99s Peak RAM: 5227.39MB | Time: 277.74s Peak RAM: 5236.08MB |
| osrm_partition | Time: 104.70s Peak RAM: 2179.76MB | Time: 106.02s Peak RAM: 2179.88MB |
| packedvector | random write:<br/>std::vector 9810.3 ms<br/>util::packed_vector 81787.4 ms<br/>slowdown: 8.33689<br/>random read:<br/>std::vector 8469.8 ms<br/>util::packed_vector 33199.1 ms<br/>slowdown: 3.91971 | random write:<br/>std::vector 9740.47 ms<br/>util::packed_vector 73379 ms<br/>slowdown: 7.53341<br/>random read:<br/>std::vector 8447.72 ms<br/>util::packed_vector 29879.3 ms<br/>slowdown: 3.53696 |
| random_match_ch |  | 1000 matches, default radius<br/>total: 78202.47ms<br/>avg: 78.20ms<br/>min: 0.48ms<br/>max: 503.92ms<br/>p99: 497.28ms<br/><br/>1000 matches, radius=10<br/>total: 276336.46ms<br/>avg: 276.34ms<br/>min: 0.60ms<br/>max: 1780.56ms<br/>p99: 1745.96ms<br/><br/>1000 matches, radius=20<br/>total: 1328821.14ms<br/>avg: 1328.82ms<br/>min: 4.38ms<br/>max: 8125.80ms<br/>p99: 8055.83ms |
| random_match_mld |  | 1000 matches, default radius<br/>total: 52138.88ms<br/>avg: 52.14ms<br/>min: 0.47ms<br/>max: 320.35ms<br/>p99: 315.54ms<br/><br/>1000 matches, radius=10<br/>total: 184917.09ms<br/>avg: 184.92ms<br/>min: 0.57ms<br/>max: 1187.55ms<br/>p99: 1165.00ms<br/><br/>1000 matches, radius=20<br/>total: 924148.77ms<br/>avg: 924.15ms<br/>min: 2.49ms<br/>max: 5294.74ms<br/>p99: 5272.57ms |
| random_nearest_ch |  | 10000 nearest, number_of_results=1<br/>total: 520.99ms<br/>avg: 0.05ms<br/>min: 0.02ms<br/>max: 0.39ms<br/>p99: 0.10ms<br/><br/>10000 nearest, number_of_results=5<br/>total: 638.95ms<br/>avg: 0.06ms<br/>min: 0.03ms<br/>max: 0.15ms<br/>p99: 0.12ms<br/><br/>10000 nearest, number_of_results=10<br/>total: 780.73ms<br/>avg: 0.08ms<br/>min: 0.04ms<br/>max: 0.17ms<br/>p99: 0.13ms |
| random_nearest_mld |  | 10000 nearest, number_of_results=1<br/>total: 519.45ms<br/>avg: 0.05ms<br/>min: 0.02ms<br/>max: 0.44ms<br/>p99: 0.11ms<br/><br/>10000 nearest, number_of_results=5<br/>total: 634.51ms<br/>avg: 0.06ms<br/>min: 0.03ms<br/>max: 0.13ms<br/>p99: 0.12ms<br/><br/>10000 nearest, number_of_results=10<br/>total: 783.69ms<br/>avg: 0.08ms<br/>min: 0.04ms<br/>max: 0.16ms<br/>p99: 0.13ms |
| random_route_ch |  | 10000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>total: 34849.64ms<br/>avg: 3.48ms<br/>min: 0.19ms<br/>max: 15.01ms<br/>p99: 5.89ms<br/><br/>10000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>total: 16845.64ms<br/>avg: 1.68ms<br/>min: 0.10ms<br/>max: 4.45ms<br/>p99: 2.95ms<br/><br/>10000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>total: 26965.10ms<br/>avg: 2.70ms<br/>min: 0.12ms<br/>max: 10.96ms<br/>p99: 5.87ms<br/><br/>10000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>total: 13132.51ms<br/>avg: 1.31ms<br/>min: 0.15ms<br/>max: 2.82ms<br/>p99: 2.05ms<br/><br/>10000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>total: 5823.38ms<br/>avg: 0.58ms<br/>min: 0.06ms<br/>max: 1.21ms<br/>p99: 0.95ms<br/><br/>10000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>total: 11388.02ms<br/>avg: 1.14ms<br/>min: 0.08ms<br/>max: 2.83ms<br/>p99: 2.20ms<br/><br/>10000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>total: 13080.85ms<br/>avg: 1.31ms<br/>min: 0.16ms<br/>max: 2.69ms<br/>p99: 2.05ms<br/><br/>10000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>total: 5845.14ms<br/>avg: 0.58ms<br/>min: 0.06ms<br/>max: 1.22ms<br/>p99: 0.95ms<br/><br/>10000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>total: 11463.23ms<br/>avg: 1.15ms<br/>min: 0.08ms<br/>max: 2.77ms<br/>p99: 2.19ms |
| random_route_mld |  | 10000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>total: 89515.91ms<br/>avg: 8.95ms<br/>min: 0.17ms<br/>max: 29.70ms<br/>p99: 17.76ms<br/><br/>10000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>total: 34238.26ms<br/>avg: 3.42ms<br/>min: 0.10ms<br/>max: 8.62ms<br/>p99: 6.05ms<br/><br/>10000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>total: 59592.65ms<br/>avg: 5.96ms<br/>min: 0.10ms<br/>max: 14.70ms<br/>p99: 11.41ms<br/><br/>10000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>total: 67406.97ms<br/>avg: 6.74ms<br/>min: 0.15ms<br/>max: 28.41ms<br/>p99: 14.35ms<br/><br/>10000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>total: 22742.97ms<br/>avg: 2.27ms<br/>min: 0.06ms<br/>max: 6.09ms<br/>p99: 4.14ms<br/><br/>10000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>total: 44386.31ms<br/>avg: 4.44ms<br/>min: 0.07ms<br/>max: 11.08ms<br/>p99: 8.12ms<br/><br/>10000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>total: 66925.50ms<br/>avg: 6.69ms<br/>min: 0.20ms<br/>max: 25.57ms<br/>p99: 14.38ms<br/><br/>10000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>total: 22920.88ms<br/>avg: 2.29ms<br/>min: 0.06ms<br/>max: 5.90ms<br/>p99: 4.29ms<br/><br/>10000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>total: 44548.52ms<br/>avg: 4.45ms<br/>min: 0.06ms<br/>max: 10.79ms<br/>p99: 8.10ms |
| random_table_ch |  | 250 tables, 3 coordinates<br/>total: 235.88ms<br/>avg: 0.94ms<br/>min: 0.61ms<br/>max: 3.91ms<br/>p99: 1.91ms<br/><br/>250 tables, 25 coordinates<br/>total: 1874.42ms<br/>avg: 7.50ms<br/>min: 6.85ms<br/>max: 8.21ms<br/>p99: 8.10ms<br/><br/>250 tables, 50 coordinates<br/>total: 3750.36ms<br/>avg: 15.00ms<br/>min: 13.86ms<br/>max: 16.05ms<br/>p99: 15.81ms<br/><br/>250 tables, 100 coordinates<br/>total: 7769.71ms<br/>avg: 31.08ms<br/>min: 29.37ms<br/>max: 34.12ms<br/>p99: 32.69ms |
| random_table_mld |  | 250 tables, 3 coordinates<br/>total: 2805.73ms<br/>avg: 11.22ms<br/>min: 7.12ms<br/>max: 18.02ms<br/>p99: 15.15ms<br/><br/>250 tables, 25 coordinates<br/>total: 25058.06ms<br/>avg: 100.23ms<br/>min: 88.52ms<br/>max: 118.71ms<br/>p99: 114.25ms<br/><br/>250 tables, 50 coordinates<br/>total: 51528.55ms<br/>avg: 206.11ms<br/>min: 190.15ms<br/>max: 228.51ms<br/>p99: 227.75ms<br/><br/>250 tables, 100 coordinates<br/>total: 109533.81ms<br/>avg: 438.14ms<br/>min: 408.95ms<br/>max: 497.78ms<br/>p99: 486.55ms |
| random_trip_ch |  | 1000 trips, 3 coordinates<br/>total: 4264.20ms<br/>avg: 4.26ms<br/>min: 1.06ms<br/>max: 13.63ms<br/>p99: 6.38ms<br/><br/>1000 trips, 4 coordinates<br/>total: 5410.67ms<br/>avg: 5.41ms<br/>min: 1.39ms<br/>max: 11.46ms<br/>p99: 7.69ms<br/><br/>1000 trips, 5 coordinates<br/>total: 6199.17ms<br/>avg: 6.20ms<br/>min: 1.59ms<br/>max: 12.47ms<br/>p99: 8.23ms |
| random_trip_mld |  | 1000 trips, 3 coordinates<br/>total: 20642.97ms<br/>avg: 20.64ms<br/>min: 7.42ms<br/>max: 37.05ms<br/>p99: 27.66ms<br/><br/>1000 trips, 4 coordinates<br/>total: 26384.84ms<br/>avg: 26.38ms<br/>min: 15.36ms<br/>max: 35.37ms<br/>p99: 32.79ms<br/><br/>1000 trips, 5 coordinates<br/>total: 32035.37ms<br/>avg: 32.04ms<br/>min: 19.99ms<br/>max: 42.84ms<br/>p99: 40.03ms |
| route_ch | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>510.256ms<br/>0.510256ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>351.894ms<br/>0.351894ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>628.455ms<br/>0.628455ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>151.514ms<br/>0.151514ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>97.5668ms<br/>0.0975668ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>133.414ms<br/>0.133414ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>151.398ms<br/>0.151398ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>97.8421ms<br/>0.0978421ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>133.048ms<br/>0.133048ms/req | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>508.457ms<br/>0.508457ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>352.12ms<br/>0.35212ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>626.2ms<br/>0.6262ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>150.082ms<br/>0.150082ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>96.9107ms<br/>0.0969107ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>132.324ms<br/>0.132324ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>150.562ms<br/>0.150562ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>97.0503ms<br/>0.0970503ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>132.339ms<br/>0.132339ms/req |
| route_mld | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>640.358ms<br/>0.640358ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>437.477ms<br/>0.437477ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>812.91ms<br/>0.81291ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>264.782ms<br/>0.264782ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>161.13ms<br/>0.16113ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>289.467ms<br/>0.289467ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>274.668ms<br/>0.274668ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>161.357ms<br/>0.161357ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>288.555ms<br/>0.288555ms/req | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>639.497ms<br/>0.639497ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>437.242ms<br/>0.437242ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>816.906ms<br/>0.816906ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>260.396ms<br/>0.260396ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>160.175ms<br/>0.160175ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>285.776ms<br/>0.285776ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>258.091ms<br/>0.258091ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>159.67ms<br/>0.15967ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>284.769ms<br/>0.284769ms/req |
| rtree | 1 result:<br/>206.687ms ->  0.0206687 ms/query<br/>10 results:<br/>241.985ms ->  0.0241985 ms/query | 1 result:<br/>207.947ms ->  0.0207947 ms/query<br/>10 results:<br/>243.336ms ->  0.0243336 ms/query |
</details>
<!-- BENCHMARK_RESULTS_END -->